requirements.txt
# pip install -r requirements.txt

# base ----------------------------------------
Cython
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.2
Pillow~=8.4.0
PyYAML>=5.3.1
scipy>=1.4.1
tensorboard>=2.2
tqdm>=4.41.0

# logging -------------------------------------
# wandb

# plotting ------------------------------------
seaborn>=0.11.0
pandas

# export --------------------------------------
# coremltools>=4.1
# onnx>=1.8.1
# scikit-learn==0.19.2  # for coreml quantization

# extras --------------------------------------
thop  # FLOPS computation
pycocotools>=2.0  # COCO mAP

torchvision~=0.10.0
yaml~=0.2.5
pyzmq~=19.0.2
flask~=2.0.2
werkzeug~=2.0.2
sqlalchemy~=1.4.22
setuptools~=58.0.4
requests~=2.27.1
pyttsx3~=2.90